hacking-session: close session in case of inactivity
If user doesn't log in to hacking session within 60 minutes,
the session is terminated. Also default behavior is changed and
the session is terminated upon logging out. When user would like
to continue hacking session the 'continue_hacking' command is provided.
Android multinode hacking session was fixed and it closes both nodes properly.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>
Change-Id: I2936067fcdc17616a2dd85bf8dd58c1d76b6f26c